% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_json.R
\name{label<-}
\alias{label<-}
\title{Set variable label}
\usage{
label(x) <- value
}
\arguments{
\item{x}{Variable}

\item{value}{Label}
}
\description{
Applies variable label as the proper type for RStudio viewing
}
