% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_json.R
\name{read_dataset_json}
\alias{read_dataset_json}
\title{Read a Dataset JSON file into R}
\usage{
read_dataset_json(json_file)
}
\arguments{
\item{json_file}{Source file location of dataset json file}
}
\value{
A tibble of the Dataset JSON contents
}
\description{
From within a Dataset JSON file, read the data into session
}
\examples{

read_dataset_json(
  url("https://raw.githubusercontent.com/lexjansen/sas-papers/master/pharmasug-2022/json/sdtm/dm.json")
)

}
