% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset_json.R
\name{type_converter}
\alias{type_converter}
\title{Convert Variable Type Generator}
\usage{
type_converter(x)
}
\arguments{
\item{x}{String indicating type in dataset json standard}
}
\description{
Function factory to return the proper function to
convert variable types
}
