% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset_json.R
\name{write_item_data}
\alias{write_item_data}
\title{Write Items to JSON file}
\usage{
write_item_data(.data)
}
\arguments{
\item{.data}{}
}
\description{
Write Items to JSON file
}
\examples{
write_item_data(df)
}
