% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset_json.R
\name{write_item_data_rows}
\alias{write_item_data_rows}
\title{Helper function to write Item Rows to JSON file}
\usage{
write_item_data_rows(.data)
}
\arguments{
\item{.data}{}
}
\description{
Helper function to write Item Rows to JSON file
}
