% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xportr-json.R
\docType{package}
\name{xportrjson-package}
\alias{xportrjson}
\alias{xportrjson-package}
\title{xportrjson: Read and Write JSON file for Clinical Trail Datasets}
\description{
A proof of concept package created for the Dataset-JSON Hackathon sponsered by CDISC Open Source Alliance. The intention is to explore R's capability to both read and write JSON files that could hold Clinical Trial Data.
}
\author{
\strong{Maintainer}: Mike Stackhouse \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Ben Straub \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Eli Miller \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Eric Simms \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
